feat(providers): extract Google Antigravity behind ProviderAdapter interface
- Add src/providers/adapter.ts contract (ProviderAdapter) and registry.ts
  with DEFAULT_PROVIDER google-antigravity + getProviderForAccount fallback
- Move OAuth, quota polling, SSE forwarding, and login flows into
  src/providers/google-antigravity/{oauth,quota,forward,login,credentials,index}.ts
- Rewrite src/login.ts as generic orchestrator delegating to adapter.runLogin()
  (Pi config sync stays Google-specific)
- Rewire rotator.ts: ensureValidToken and quota polling now go through the
  adapter; remove inlined fetchQuota/extractQuotas/getOAuthClientConfig code
- Move KICKSTART_MODEL_FOR_QUOTA_POOL to types.ts for shared access
- Move compat/translators.ts into the Google provider; compat.ts re-exports
  unchanged; fix imports in compression/* and tests
- AccountConfig gains optional provider field (legacy configs default to
  google-antigravity via registry fallback)
