+++ src/compression/rtk/index.ts
@@ -1,4 +1,4 @@
-import type { ChatMessage } from "../../compat/translators.js";
+import type { ChatMessage } from "../../providers/google-antigravity/translators.js";
 import {
   extractPreservedBlocks,
   restorePreservedBlocks,
