+++ src/providers/registry.ts
@@ -1,0 +1,51 @@
+// Provider registry: maps provider ids to their adapters.
+//
+// To add a new provider:
+//   1. Implement ProviderAdapter in src/providers/<id>/index.ts
+//   2. Add an entry to PROVIDERS below
+//   3. Extend the AccountConfig union in src/types.ts with your account shape
+//   4. Wire your login flow through `rotator login --provider <id>`
+
+import type { ProviderAdapter } from "./adapter.js";
+import { googleAntigravityAdapter } from "./google-antigravity/index.js";
+import { logger } from "../logger.js";
+
+export const DEFAULT_PROVIDER = "google-antigravity";
+
+const PROVIDERS: Record<string, ProviderAdapter> = {
+  "google-antigravity": googleAntigravityAdapter,
+};
+
+export function getProviderAdapter(providerId: string): ProviderAdapter {
+  const adapter = PROVIDERS[providerId];
+  if (!adapter) {
+    throw new Error(
+      `Unknown provider "${providerId}". Available: ${Object.keys(PROVIDERS).join(", ")}`,
+    );
+  }
+  return adapter;
+}
+
+export function listProviders(): ProviderAdapter[] {
+  return Object.values(PROVIDERS);
+}
+
+export function isKnownProvider(providerId: string): boolean {
+  return Object.prototype.hasOwnProperty.call(PROVIDERS, providerId);
+}
+
+export function getProviderForAccount(
+  account: { provider?: string },
+): ProviderAdapter {
+  const id = account.provider ?? DEFAULT_PROVIDER;
+  try {
+    return getProviderAdapter(id);
+  } catch {
+    logger
+      .child("registry")
+      .warn(
+        `Account references unknown provider "${id}", falling back to ${DEFAULT_PROVIDER}; check config`,
+      );
+    return getProviderAdapter(DEFAULT_PROVIDER);
+  }
+}
