+++ src/proxy.ts
@@ -12,16 +12,19 @@ import {
 } from "node:http";
 import { Readable } from "node:stream";
 import {
-  ANTIGRAVITY_ENDPOINTS,
-  REQUEST_CLIENT_METADATA,
-  REQUEST_GOOG_API_CLIENT,
-  REQUEST_USER_AGENT,
-  applyModelAlias,
   resolveQuotaModelKey,
   resolveDisplayModelKey,
 } from "./types.js";
 import type { AccountRuntime } from "./types.js";
 import type { AccountRotator } from "./rotator.js";
+import {
+  forwardRequest,
+  SseEventAccumulator,
+  extractUsageFromSseEvent,
+  getBenchmarkSpec,
+  GOOGLE_BENCHMARK_CONSTANTS,
+} from "./providers/google-antigravity/forward.js";
+export { forwardRequest, SseEventAccumulator, extractUsageFromSseEvent };
 import {
   serveDashboard,
   serveStatusApi,
@@ -613,200 +616,6 @@ function isFetchTransportError(err: unknown): boolean {
   return false;
 }
 
-/** Max bytes kept in the SSE event buffer. A single event is rarely >1MB;
- *  if it is, we keep the last 1MB which is still enough to find usage. */
-const SSE_EVENT_BUFFER_MAX = 1024 * 1024;
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-/** Recursively search a parsed JSON value for the first usage block we recognise.
- *  Supports Gemini (usageMetadata), OpenAI (usage with prompt_tokens/completion_tokens),
- *  and Anthropic (usage with input_tokens/output_tokens). */
-function findUsageInJson(
-  value: unknown,
-): { inputTokens: number; outputTokens: number } | null {
-  if (!isRecord(value)) return null;
-  // Gemini format
-  const gemini = value.usageMetadata;
-  if (isRecord(gemini)) {
-    const input =
-      typeof gemini.promptTokenCount === "number" ? gemini.promptTokenCount : 0;
-    const output =
-      typeof gemini.candidatesTokenCount === "number"
-        ? gemini.candidatesTokenCount
-        : 0;
-    if (input > 0 || output > 0)
-      return { inputTokens: input, outputTokens: output };
-  }
-  // OpenAI / Anthropic format
-  const usage = value.usage;
-  if (isRecord(usage)) {
-    const input =
-      typeof usage.prompt_tokens === "number"
-        ? usage.prompt_tokens
-        : typeof usage.input_tokens === "number"
-          ? usage.input_tokens
-          : 0;
-    const output =
-      typeof usage.completion_tokens === "number"
-        ? usage.completion_tokens
-        : typeof usage.output_tokens === "number"
-          ? usage.output_tokens
-          : 0;
-    if (input > 0 || output > 0)
-      return { inputTokens: input, outputTokens: output };
-  }
-  // Recurse into common nesting locations.
-  for (const key of ["candidates", "output", "response", "message"]) {
-    const child = value[key];
-    if (Array.isArray(child)) {
-      for (const item of child) {
-        const found = findUsageInJson(item);
-        if (found) return found;
-      }
-    } else if (isRecord(child)) {
-      const found = findUsageInJson(child);
-      if (found) return found;
-    }
-  }
-  return null;
-}
-
-/** Extract usage from a single complete SSE event (one or more `data:` lines
- *  separated by newlines, terminated by a blank line). The last successful
- *  extraction wins (callers should stop scanning once they find usage). */
-export function extractUsageFromSseEvent(
-  eventText: string,
-): { inputTokens: number; outputTokens: number } | null {
-  const dataLines: string[] = [];
-  for (const raw of eventText.split("\n")) {
-    if (raw.startsWith("data:")) {
-      dataLines.push(raw.slice(5).trim());
-    }
-  }
-  if (dataLines.length === 0) return null;
-  const payload = dataLines.join("\n");
-  if (payload === "[DONE]" || payload === "") return null;
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(payload);
-  } catch {
-    // Fall back to regex on the raw event text. This handles non-standard
-    // streams that don't quite produce valid JSON per event.
-    const fallback = regexExtractUsage(payload);
-    return fallback;
-  }
-  return findUsageInJson(parsed);
-}
-
-/** Last-resort regex extraction for streams that don't yield parseable JSON. */
-function regexExtractUsage(
-  buffer: string,
-): { inputTokens: number; outputTokens: number } | null {
-  try {
-    const patterns = [
-      /"promptTokenCount"\s*:\s*(\d+).*?"candidatesTokenCount"\s*:\s*(\d+)/s,
-      /"input_tokens"\s*:\s*(\d+).*?"output_tokens"\s*:\s*(\d+)/s,
-    ];
-    for (const pattern of patterns) {
-      const match = buffer.match(pattern);
-      if (match) {
-        return {
-          inputTokens: parseInt(match[1], 10),
-          outputTokens: parseInt(match[2], 10),
-        };
-      }
-    }
-  } catch {
-    /* extraction failed */
-  }
-  return null;
-}
-
-/** State for the SSE event accumulator used by streamResponseBody. */
-class SseEventAccumulator {
-  private buffer = "";
-  private accumulatedText = "";
-  private readonly maxBytes: number;
-  constructor(maxBytes: number = SSE_EVENT_BUFFER_MAX) {
-    this.maxBytes = maxBytes;
-  }
-
-  /** Append a chunk, return any usage extracted from newly-completed events. */
-  append(
-    chunkText: string,
-  ): { inputTokens: number; outputTokens: number } | null {
-    this.buffer += chunkText;
-    if (this.buffer.length > this.maxBytes) {
-      this.buffer = this.buffer.slice(-this.maxBytes);
-    }
-    let extracted: { inputTokens: number; outputTokens: number } | null = null;
-    let boundary = this.buffer.indexOf("\n\n");
-    while (boundary !== -1) {
-      const eventText = this.buffer.slice(0, boundary);
-      this.buffer = this.buffer.slice(boundary + 2);
-      const usage = extractUsageFromSseEvent(eventText);
-      if (usage && !extracted) extracted = usage;
-      const text = extractTextFromSseEvent(eventText);
-      if (text && this.accumulatedText.length < 100000) {
-        this.accumulatedText += text;
-      }
-      boundary = this.buffer.indexOf("\n\n");
-    }
-    return extracted;
-  }
-
-  getText(): string {
-    return this.accumulatedText;
-  }
-
-  /** Flush any partial event at end-of-stream. */
-  final(): { inputTokens: number; outputTokens: number } | null {
-    if (!this.buffer) return null;
-    const usage = extractUsageFromSseEvent(this.buffer);
-    const text = extractTextFromSseEvent(this.buffer);
-    if (text && this.accumulatedText.length < 100000) {
-      this.accumulatedText += text;
-    }
-    this.buffer = "";
-    return usage;
-  }
-}
-
-function extractTextFromSseEvent(eventText: string): string | null {
-  const dataLines: string[] = [];
-  for (const raw of eventText.split("\n")) {
-    if (raw.startsWith("data:")) {
-      dataLines.push(raw.slice(5).trim());
-    }
-  }
-  if (dataLines.length === 0) return null;
-  const payload = dataLines.join("\n");
-  if (payload === "[DONE]" || payload === "") return null;
-  try {
-    const parsed = JSON.parse(payload);
-    if (parsed && typeof parsed === "object") {
-      const p = parsed as Record<string, unknown>;
-      if (Array.isArray(p.candidates) && p.candidates[0] && typeof p.candidates[0] === "object") {
-        const cand = p.candidates[0] as Record<string, unknown>;
-        if (cand.content && typeof cand.content === "object") {
-          const content = cand.content as Record<string, unknown>;
-          if (Array.isArray(content.parts)) {
-            return content.parts
-              .map((part) => (part && typeof part === "object" && typeof (part as Record<string, unknown>).text === "string" ? (part as Record<string, unknown>).text : ""))
-              .join("");
-          }
-        }
-      }
-    }
-  } catch {
-    /* ignore */
-  }
-  return null;
-}
-
 async function readJsonRequest(req: IncomingMessage): Promise<unknown> {
   const body = await readLimitedBody(req);
   return body.length === 0 ? {} : JSON.parse(body.toString("utf-8"));
@@ -982,167 +791,7 @@ async function streamResponseBody(
   }
 
   return usage;
-}
-
-/**
- * Forward a request to the real Antigravity endpoint with credential swapping.
- */
-export async function forwardRequest(
-  account: AccountRuntime,
-  body: RequestBody,
-  originalHeaders: Record<string, string>,
-  signal?: AbortSignal,
-): Promise<ForwardedResponse> {
-  // Swap credentials
-  body.project = account.config.projectId;
-
-  // Map internal display/compat names to Google upstream names (single source
-  // of truth: src/types.ts:applyModelAlias)
-  body.model = applyModelAlias(body.model);
-
-  const { displayModel: _displayModel, ...bodyToForward } = body;
-  const requestBody = JSON.stringify(bodyToForward);
-
-  // Build headers: keep originals but swap Authorization
-  const forwardHeaders: Record<string, string> = {
-    ...originalHeaders,
-    "Content-Type": "application/json",
-    Accept: "text/event-stream",
-  };
-  // Remove original authorization (any case), provider-set headers, and
-  // hop-by-hop headers per RFC 7230 §6.1. The hop-by-hop list prevents
-  // leaking client IP (X-Forwarded-For) and prevents IP spoofing in
-  // upstream logs (Via).
-  const HOP_BY_HOP = new Set([
-    "connection",
-    "keep-alive",
-    "proxy-authenticate",
-    "proxy-authorization",
-    "te",
-    "trailers",
-    "transfer-encoding",
-    "upgrade",
-    // Forwarding / proxying artefacts that should never reach the upstream
-    "x-forwarded-for",
-    "x-forwarded-host",
-    "x-forwarded-proto",
-    "x-forwarded-port",
-    "x-real-ip",
-    "forwarded",
-    "via",
-  ]);
-  for (const key of Object.keys(forwardHeaders)) {
-    const lowerKey = key.toLowerCase();
-    if (
-      HOP_BY_HOP.has(lowerKey) ||
-      lowerKey === "authorization" ||
-      lowerKey === "user-agent" ||
-      lowerKey === "x-goog-api-client" ||
-      lowerKey === "client-metadata"
-    ) {
-      delete forwardHeaders[key];
-    }
-  }
-  forwardHeaders["Authorization"] = `Bearer ${account.accessToken}`;
-  forwardHeaders["User-Agent"] = REQUEST_USER_AGENT;
-  forwardHeaders["X-Goog-Api-Client"] = REQUEST_GOOG_API_CLIENT;
-  forwardHeaders["Client-Metadata"] = REQUEST_CLIENT_METADATA;
-  // Claude models on Cloud Code Assist (Antigravity) require this beta header to
-  // return interleaved thinking blocks. Mirrors pi-mono's needsClaudeThinkingBetaHeader.
-  if (/^claude-/i.test(body.model)) {
-    forwardHeaders["anthropic-beta"] = "interleaved-thinking-2025-05-14";
-  }
-  delete forwardHeaders["host"];
-  delete forwardHeaders["connection"];
-  delete forwardHeaders["transfer-encoding"];
-  delete forwardHeaders["content-length"];
-
-  // Try endpoints with cascade on 401/403/404
-  for (
-    let endpointIdx = 0;
-    endpointIdx < ANTIGRAVITY_ENDPOINTS.length;
-    endpointIdx++
-  ) {
-    const endpoint = ANTIGRAVITY_ENDPOINTS[endpointIdx];
-    const url = `${endpoint}/v1internal:streamGenerateContent?alt=sse`;
-    const isProd = endpointIdx === ANTIGRAVITY_ENDPOINTS.length - 1;
-
-    try {
-      const controller = !isProd ? new AbortController() : undefined;
-      const timeout = controller
-        ? setTimeout(() => controller.abort(), 10_000)
-        : undefined;
-      const requestSignal = controller?.signal && signal
-        ? AbortSignal.any([controller.signal, signal])
-        : signal ?? controller?.signal;
-
-      const response = await fetch(url, {
-        method: "POST",
-        headers: forwardHeaders,
-        body: requestBody,
-        signal: requestSignal,
-      });
-      if (timeout) clearTimeout(timeout);
-
-      if (
-        (response.status === 401 ||
-          response.status === 403 ||
-          response.status === 404) &&
-        endpointIdx < ANTIGRAVITY_ENDPOINTS.length - 1
-      ) {
-        log(`Endpoint ${endpoint} returned ${response.status}, cascading...`);
-        response.text().catch(() => {});
-        continue;
-      }
-
-      return { response, endpoint };
-    } catch (err) {
-      if (endpointIdx < ANTIGRAVITY_ENDPOINTS.length - 1) {
-        log(
-          `Endpoint ${endpoint} failed: ${err instanceof Error ? err.message : err}, cascading...`,
-        );
-        continue;
-      }
-      throw err;
-    }
-  }
-
-  throw new Error("All endpoints failed");
-}
-
-const BENCHMARK_MODEL = "gemini-3-flash";
-const BENCHMARK_PROMPT = "Reply with exactly: OK";
-const BENCHMARK_MAX_OUTPUT_TOKENS = 16;
-const BENCHMARK_TIMEOUT_MS = 30_000;
-
-function benchmarkRequestBody(account: AccountRuntime): RequestBody {
-  return {
-    project: account.config.projectId,
-    model: BENCHMARK_MODEL,
-    request: {
-      contents: [
-        {
-          role: "user",
-          parts: [{ text: BENCHMARK_PROMPT }],
-        },
-      ],
-      generationConfig: {
-        maxOutputTokens: BENCHMARK_MAX_OUTPUT_TOKENS,
-      },
-    },
-  };
-}
-
-function benchmarkUsage(raw: string): { outputTokens: number } | null {
-  let outputTokens = 0;
-  for (const event of raw.split(/\r?\n\r?\n/)) {
-    const usage = extractUsageFromSseEvent(event);
-    if (usage) outputTokens = usage.outputTokens;
-  }
-  return outputTokens > 0 ? { outputTokens } : null;
-}
-
-function benchmarkFailure(
+}function benchmarkFailure(
   account: AccountRuntime,
   startedAt: number,
   error: string,
@@ -1182,11 +831,12 @@ export async function benchmarkAccount(
     };
   }
 
-  rotator.startRequest(account, BENCHMARK_MODEL);
+  const benchmarkSpec = getBenchmarkSpec(account);
+  rotator.startRequest(account, GOOGLE_BENCHMARK_CONSTANTS.model);
   const timeoutController = new AbortController();
   const timeout = setTimeout(
     () => timeoutController.abort(new Error("Benchmark timeout")),
-    BENCHMARK_TIMEOUT_MS,
+    GOOGLE_BENCHMARK_CONSTANTS.timeoutMs,
   );
   const signal = sharedSignal
     ? AbortSignal.any([timeoutController.signal, sharedSignal])
@@ -1197,7 +847,7 @@ export async function benchmarkAccount(
     await rotator.ensureValidToken(account);
     const forwarded = await forwardRequest(
       account,
-      benchmarkRequestBody(account),
+      benchmarkSpec.body,
       {},
       signal,
     );
@@ -1213,7 +863,7 @@ export async function benchmarkAccount(
       );
     }
 
-    const usage = benchmarkUsage(raw);
+    const usage = benchmarkSpec.parseUsage(raw);
     const outputTokens = usage?.outputTokens ?? Math.max(1, Math.ceil(raw.length / 4));
     return {
       account: label,
@@ -1227,7 +877,7 @@ export async function benchmarkAccount(
     return benchmarkFailure(account, startedAt, formatError(err), ttfbMs);
   } finally {
     clearTimeout(timeout);
-    rotator.finishRequest(account, BENCHMARK_MODEL);
+    rotator.finishRequest(account, GOOGLE_BENCHMARK_CONSTANTS.model);
   }
 }
 
@@ -1311,7 +961,7 @@ export async function serveBenchmarkApi(
     writeBenchmarkEvent(res, {
       type: "start",
       total: accounts.length,
-      model: BENCHMARK_MODEL,
+      model: GOOGLE_BENCHMARK_CONSTANTS.model,
     });
     await Promise.all(
       accounts.map(async (account) => {
